Refactor scheduling system
