+++ src/main/kotlin/me/snoty/backend/scheduling/NodeScheduler.kt
@@ -1,0 +1,7 @@
+package me.snoty.backend.scheduling
+
+import me.snoty.integration.common.wiring.Node
+
+interface NodeScheduler {
+	fun schedule(node: Node)
+}
