+++ src/main/kotlin/me/snoty/backend/scheduling/impl/jobrunr/JobRunrConfigurer.kt
@@ -1,15 +1,11 @@
-package me.snoty.backend.scheduling
+package me.snoty.backend.scheduling.impl.jobrunr
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.registerKotlinModule
 import com.mongodb.client.MongoClient
 import io.micrometer.core.instrument.MeterRegistry
 import me.snoty.backend.database.mongo.MONGO_DB_NAME
-import me.snoty.backend.integration.flow.logging.NodeLogService
-import me.snoty.backend.scheduling.node.NodeJobHandler
-import me.snoty.integration.common.config.NodeService
-import me.snoty.integration.common.wiring.flow.FlowService
-import me.snoty.integration.common.wiring.node.NodeRegistry
+import me.snoty.backend.scheduling.impl.jobrunr.node.JobRunrNodeJobHandler
 import org.jobrunr.configuration.JobRunr
 import org.jobrunr.configuration.JobRunrMicroMeterIntegration
 import org.jobrunr.dashboard.JobRunrDashboardWebServerConfiguration
@@ -17,17 +13,15 @@ import org.jobrunr.server.JobActivator
 import org.jobrunr.storage.StorageProviderUtils
 import org.jobrunr.storage.nosql.mongo.MongoDBStorageProvider
 import org.jobrunr.utils.mapper.jackson.JacksonJsonMapper
+import org.koin.core.annotation.Single
 
-object JobRunrConfigurer {
-	fun configure(
-		mongoClient: MongoClient,
-		nodeRegistry: NodeRegistry,
-		nodeService: NodeService,
-		flowService: FlowService,
-		nodeLogService: NodeLogService,
-		meterRegistry: MeterRegistry,
-	) {
-		val jobHandler = NodeJobHandler(nodeRegistry, nodeService, flowService, nodeLogService)
+@Single
+class JobRunrConfigurer(
+	private val mongoClient: MongoClient,
+	private val meterRegistry: MeterRegistry,
+	private val jobHandler: JobRunrNodeJobHandler,
+) {
+	fun configure() {
 		JobRunr.configure()
 			.useJsonMapper(JacksonJsonMapper(ObjectMapper().registerKotlinModule()))
 			.useStorageProvider(
@@ -42,7 +36,7 @@ object JobRunrConfigurer {
 				override fun <T : Any> activateJob(type: Class<T>): T? {
 					@Suppress("UNCHECKED_CAST")
 					return when (type) {
-						NodeJobHandler::class.java -> jobHandler as T
+						JobRunrNodeJobHandler::class.java -> jobHandler as T
 						else -> null
 					}
 				}
