+++ src/main/kotlin/me/snoty/backend/scheduling/impl/jobrunr/node/JobRunrNodeJobRequest.kt
@@ -1,0 +1,10 @@
+package me.snoty.backend.scheduling.impl.jobrunr.node
+
+import me.snoty.backend.integration.config.flow.NodeId
+import me.snoty.backend.scheduling.JobRequest
+
+data class JobRunrNodeJobRequest(
+	val nodeId: NodeId
+) : JobRequest {
+	override fun getJobRequestHandler() = JobRunrNodeJobHandler::class.java
+}
