+++ src/test/kotlin/me/snoty/backend/test/RestTest.kt
@@ -2,9 +2,7 @@ package me.snoty.backend.test
 
 import io.ktor.server.routing.*
 import io.ktor.server.testing.*
-import io.mockk.mockk
 import me.snoty.backend.config.Config
-import me.snoty.backend.server.routing.addResources
 import me.snoty.backend.server.plugins.configureRouting
 import me.snoty.backend.server.plugins.configureSecurity
 import me.snoty.backend.server.plugins.configureSerialization
@@ -20,7 +18,6 @@ fun ktorApplicationTest(
 			configureSerialization(BaseSnotyJson)
 			configureSecurity(config)
 			configureRouting(config)
-			addResources(mockk(relaxed = true))
 		}
 
 		routing {
